some fix + test
