+++ src/components/place-cards/place-card-list.tsx
@@ -1,6 +1,6 @@
 import { useAppSelector } from '../../hooks';
 import {Offers} from '../../types/offer';
-import PlaceCard from '../place-card/place-card';
+import PlaceCard from './place-card';
 import sortingFunction from '../../util';
 import { memo } from 'react';
 import { getSortingOption } from '../../store/other-process/selectors';
@@ -10,7 +10,7 @@ type PlaceCardList = {
   typeOfList: 'typical' | 'near';
 };
 
-function PlacesCardList({offers, typeOfList}: PlaceCardList): JSX.Element {
+function PlacesCardList({offers, typeOfList}: PlaceCardList){
   const currnetSortOptions = useAppSelector(getSortingOption);
 
   return (
