+++ src/components/use-map/use-map.tsx
@@ -1,6 +1,6 @@
 import {useEffect, useState, MutableRefObject, useRef} from 'react';
 import {Map, TileLayer} from 'leaflet';
-import {City} from '../types/offer';
+import { City } from '../../types/offer';
 
 
 function useMap(
