+++ src/service/token.ts
@@ -1,16 +1,16 @@
-const AUTH_TOKEN_KEY_NAME = 'six-cities-token';
+const TOKEN_KEY_NAME = 'six-cities-token';
 
 export type Token = string;
 
 export const getToken = (): Token => {
-  const token = localStorage.getItem(AUTH_TOKEN_KEY_NAME);
+  const token = localStorage.getItem(TOKEN_KEY_NAME);
   return token ?? '';
 };
 
 export const saveToken = (token: Token): void => {
-  localStorage.setItem(AUTH_TOKEN_KEY_NAME, token);
+  localStorage.setItem(TOKEN_KEY_NAME, token);
 };
 
 export const dropToken = (): void => {
-  localStorage.removeItem(AUTH_TOKEN_KEY_NAME);
+  localStorage.removeItem(TOKEN_KEY_NAME);
 };
